#11 - 회원 패키지 정리
회원(member) 관련 코드 - 패키지 정리
