--- src/main/java/com/example/exbasicspring/domain/Grade.java
@@ -1,6 +1,0 @@
-package com.example.exbasicspring.domain;
-
-public enum Grade {
-    BASIC,
-    VIP
-}
