+++ src/main/java/com/example/exbasicspring/member/domain/Grade.java
@@ -1,0 +1,6 @@
+package com.example.exbasicspring.member.domain;
+
+public enum Grade {
+    BASIC,
+    VIP
+}
