+++ src/main/java/com/example/exbasicspring/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.example.exbasicspring.member.repository;
+
+import com.example.exbasicspring.member.domain.Member;
+
+public interface MemberRepository {
+
+    void save(Member member);
+
+    Member findById(Long id);
+
+}
