+++ src/main/java/com/example/exbasicspring/member/service/MemberService.java
@@ -1,0 +1,11 @@
+package com.example.exbasicspring.member.service;
+
+import com.example.exbasicspring.member.domain.Member;
+
+public interface MemberService {
+
+    void join(Member member);
+
+    Member findMember(Long id);
+
+}
