--- src/main/java/com/example/exbasicspring/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.example.exbasicspring.repository;
-
-import com.example.exbasicspring.domain.Member;
-
-public interface MemberRepository {
-
-    void save(Member member);
-
-    Member findById(Long id);
-
-}
