--- src/main/java/com/example/exbasicspring/service/MemberService.java
@@ -1,11 +1,0 @@
-package com.example.exbasicspring.service;
-
-import com.example.exbasicspring.domain.Member;
-
-public interface MemberService {
-
-    void join(Member member);
-
-    Member findMember(Long id);
-
-}
